Migrate to solidjs
