+++ Cargo.toml
@@ -10,7 +10,6 @@ edition = "2021"
 mio = "0.8.11"
 h2 = "0.4.4"
 anyhow = "1.0.79"
-askama = "0.12.1"
 axum = "0.7.3"
 blowfish = "0.9.1"
 chrono = "0.4.31"
@@ -27,7 +26,6 @@ lazy_static = "1.4.0"
 log = "0.4.20"
 phf = { version = "0.11.2", features = ["macros"] }
 rand = "0.8.5"
-rayon = "1.8.0"
 regex = "1.10.2"
 reqwest = { version = "0.11.23", default-features = false, features = [
   "rustls-tls",
