--- templates/paste.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<script type="module">
-	window.PASTE_ID = "{{ paste }}";
-</script>
-
-<head>
-	<meta charset="UTF-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<!--<link rel="stylesheet" href="/assets/paste.css" />-->
-	<link rel="apple-touch-icon" sizes="180x180" href="/assets/favicon/apple-touch-icon.png">
-	<link rel="icon" type="image/png" sizes="32x32" href="/assets/favicon/favicon-32x32.png">
-	<link rel="icon" type="image/png" sizes="16x16" href="/assets/favicon/favicon-16x16.png">
-	<link rel="manifest" href="/site.webmanifest">
-    <title>Pokebin</title>
-</head>
-
-<body>
-	<div id="app"></div>
-	<!--<script type="module" crossorigin src="/static/js/paste.js"></script>-->
-</body>
-
-</html>
