+++ web/about.html
@@ -14,12 +14,12 @@
 	body {
 		background-color: black;
 		overflow: hidden;
+        color: white;
 	}
 </style>
 
 <body>
 	<div id="app"></div>
-	<script type="module" src="/src/about.ts"></script>
 </body>
 
 </html>
