--- web/biome.json
@@ -1,18 +1,0 @@
-{
-	"$schema": "https://biomejs.dev/schemas/1.4.1/schema.json",
-	"organizeImports": {
-		"enabled": true
-	},
-	"linter": {
-		"enabled": true,
-		"rules": {
-			"recommended": true
-		}
-	},
-
-	"formatter": {
-		"enabled": true,
-		"indentStyle": "space",
-		"indentWidth": 4
-	}
-}
