+++ web/index.html
@@ -13,7 +13,6 @@
 
 <body>
 	<div id="app"></div>
-	<script type="module" src="/src/main.ts"></script>
 </body>
 
 </html>
