+++ web/package.json
@@ -1,32 +1,28 @@
 {
-  "name": "web",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "watch": "vite build --watch",
-    "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.4.2",
-    "@tsconfig/svelte": "^5.0.0",
-    "autoprefixer": "^10.4.16",
-    "postcss": "^8.4.33",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.6",
-    "tailwindcss": "^3.4.1",
-    "tslib": "^2.6.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.4.5"
-  },
-  "dependencies": {
-    "@types/dompurify": "^3.0.5",
-    "axios": "^1.6.5",
-    "bakutils-catcher": "^2.0.0",
-    "dompurify": "^3.0.8",
-    "ts-pattern": "^5.0.8"
-  }
+    "name": "solid",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+        "dev": "rsbuild serve",
+        "build": "rsbuild build"
+    },
+    "dependencies": {
+        "bakutils-catcher": "^3.6.1",
+        "dompurify": "^3.1.6",
+        "solid-js": "^1.8.17",
+        "ts-pattern": "^5.2.0"
+    },
+    "devDependencies": {
+        "@rsbuild/core": "^1.0.1-beta.3",
+        "@rsbuild/plugin-babel": "^1.0.1-beta.3",
+        "@rsbuild/plugin-solid": "^1.0.1-beta.3",
+        "@rspack/cli": "^0.7.5",
+        "@rspack/core": "^0.7.5",
+        "@types/dompurify": "^3.0.5",
+        "autoprefixer": "^10.4.19",
+        "postcss": "^8.4.39",
+        "tailwindcss": "^3.4.4",
+        "typescript": "^5.2.2"
+    }
 }
