--- web/solid/about.html
@@ -1,25 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>Pokebin</title>
-	<link rel="apple-touch-icon" sizes="180x180" href="/assets/favicon/apple-touch-icon.png">
-	<link rel="icon" type="image/png" sizes="32x32" href="/assets/favicon/favicon-32x32.png">
-	<link rel="icon" type="image/png" sizes="16x16" href="/assets/favicon/favicon-16x16.png">
-	<link rel="manifest" href="/site.webmanifest">
-</head>
-<style>
-	body {
-		background-color: black;
-		overflow: hidden;
-        color: white;
-	}
-</style>
-
-<body>
-	<div id="app"></div>
-</body>
-
-</html>
