--- web/solid/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "solid",
-    "private": true,
-    "version": "0.0.0",
-    "type": "module",
-    "scripts": {
-        "dev": "rsbuild serve",
-        "build": "rsbuild build"
-    },
-    "dependencies": {
-        "bakutils-catcher": "^3.6.1",
-        "dompurify": "^3.1.6",
-        "solid-js": "^1.8.17",
-        "ts-pattern": "^5.2.0"
-    },
-    "devDependencies": {
-        "@rsbuild/core": "^1.0.1-beta.3",
-        "@rsbuild/plugin-babel": "^1.0.1-beta.3",
-        "@rsbuild/plugin-solid": "^1.0.1-beta.3",
-        "@rspack/cli": "^0.7.5",
-        "@rspack/core": "^0.7.5",
-        "autoprefixer": "^10.4.19",
-        "postcss": "^8.4.39",
-        "tailwindcss": "^3.4.4",
-        "typescript": "^5.2.2"
-    }
-}
