--- web/solid/src/paste.css
@@ -1,356 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-.toggle-notes {
-    background: #333;
-    color: white;
-    border: none;
-    padding: 10px;
-    cursor: pointer;
-    user-select: none !important;
-    border-radius: 3px;
-}
-
-p {
-	word-wrap: break-word;
-}
-
-@media (max-width: 399px) {
-	article .paste {
-		font-size: 13px !important;
-		min-width: 100% !important;
-		width: 100% !important;
-	}
-}
-
-
-/* Mobile styles */
-@media (max-width: 1158px) {
-
-	.extra {
-		flex-direction: column;
-		gap: 10px;
-	}
-
-	body {
-		font-size: 100%;
-		/* Smaller font size for mobile */
-	}
-
-	.content-container {
-		display: flex;
-		flex-direction: column;
-		/* side-content above main */
-	}
-
-	.side-content {
-		width: 100% !important;
-		flex-direction: column !important;
-	}
-
-    article>.paste {
-        max-width: 350px;
-    }
-
-}
-
-img.img-pokemon {
-    max-width: 180px;
-    /* max-width: 200px; */
-    min-width: 150px;
-    width: 15vw;
-    height: auto;
-    margin-left: 15px;
-}
-
-/* Desktop and larger tablets */
-@media (min-width: 1159px) {
-
-	.extra {
-		flex-direction: column;
-	}
-
-	body {
-        font-size: 100%;
-		/* Regular font size for larger devices */
-	}
-
-	.content-container {
-		display: flex;
-		flex-direction: column;
-		/* main above side-content */
-	}
-
-    article>.paste {
-        max-width: 450px;
-    }
-    /* img.img-pokemon { */
-    /*     max-width: 300px; */
-    /*     min-width: 150px; */
-    /*     width: 14vw; */
-    /*  height: auto; */
-    /* } */
-
-}
-
-/* img.img-pokemon { */
-/* 	min-width: 150px; */
-/* 	height: auto; */
-/* 	max-width: 300px; */
-/*     width: 13vw; */
-/* } */
-
-body {
-	margin: 0;
-	background-color: black !important;
-    /*@apply bg-zinc-900;*/
-	color: white;
-	/* font: */
-	/* 	12px "M+ 1c", */
-	/* 	sans-serif; */
-}
-
-main,
-.side-content {
-	margin: 0;
-}
-
-.side-content {
-	padding-top: 20px !important;
-}
-
-.extra {
-	width: 200px;
-	user-select: none !important;
-    margin-top: 10px;
-}
-
-/*.side-content {*/
-/*	display: flex;*/
-	/*flex-direction: row;*/
-	/*flex-direction: column;*/
-	/*align-items: flex-start;*/
-	/* gap: 20px; */
-/*}*/
-
-* {
-	min-width: 0;
-}
-
-.type-normal {
-	color: #a8a878;
-}
-
-.type-fighting {
-	color: #c03028;
-}
-
-.type-flying {
-	color: #a890f0;
-}
-
-.type-poison {
-	color: #a040a0;
-}
-
-.type-ground {
-	color: #e0c068;
-}
-
-.type-rock {
-	color: #b8a038;
-}
-
-.type-bug {
-	color: #a8b820;
-}
-
-.type-ghost {
-	color: #705898;
-}
-
-.type-steel {
-	color: #b8b8d0;
-}
-
-.type-fire {
-	color: #f08030;
-}
-
-.type-water {
-	color: #6890f0;
-}
-
-.type-grass {
-	color: #78c850;
-}
-
-.type-electric {
-	color: #f8d030;
-}
-
-.type-psychic {
-	color: #f85888;
-}
-
-.type-ice {
-	color: #98d8d8;
-}
-
-.type-dragon {
-	color: #7038f8;
-}
-
-.type-dark {
-	color: #705848;
-}
-
-.type-fairy {
-	color: #ee99ac;
-}
-
-.gender-m {
-	@apply text-sky-500;
-}
-
-.gender-f {
-	color: #e82010;
-}
-
-.attr {
-	color: #a0a0a0;
-}
-
-.stat-hp {
-	color: #ff0000;
-}
-
-.stat-atk {
-	color: #f08030;
-}
-
-.stat-def {
-	color: #f8d030;
-}
-
-.stat-spa {
-	color: #6890f0;
-}
-
-.stat-spd {
-	color: #78c850;
-}
-
-.stat-spe {
-	color: #f85888;
-}
-
-main,
-.side-content {
-	margin: 0
-}
-
-article {
-	float: left;
-}
-
-article:nth-child(odd) {
-	clear: left;
-}
-
-article>.paste {
-	min-width: 335px;
-    width: 25vw;
-    /*max-width: 300px;*/
-    margin: 2px 5px 0;
-	/*font: 15px;*/
-    font-size: 1rem;
-    line-height: 1.3rem;
-    /*font-family: monospace;*/
-    font-family: 'Ubuntu Mono', monospace;
-    font-weight: 400;
-    font-style: normal;
-	/*line-height: 1.45;*/
-}
-
-article {
-	display: flex;
-	flex-flow: row wrap;
-	align-items: flex-start;
-	margin: 0;
-}
-
-div.img {
-	position: relative;
-	margin-left: 0px;
-	user-select: none;
-	-webkit-user-select: none;
-	-moz-user-select: none;
-	-ms-user-select: none;
-}
-
-span.img-item {
-	position: absolute;
-	right: 0;
-	bottom: 0;
-	width: 48px;
-	/* Change this to the desired width */
-	height: 48px;
-	/* Change this to the desired height */
-	/* background-size: 344px 1032px; */
-	/* Adjust the background size to 'auto' */
-}
-
-
-#wrapper {
-	user-select: none;
-	width: 100%;
-	/* Adjust the width as necessary, could be 100% or a specific value like 35vw */
-	max-width: 35vw;
-	/* Prevents the container from exceeding the bounds of its parent */
-	height: 25px;
-	/* Fixed height as you've already specified */
-	max-height: 25px;
-	/* Ensures container does not grow taller than this */
-	overflow: hidden;
-	/* Adds scrollbars if content exceeds container size */
-	margin: 0 auto;
-	/* Centers the container if within a block that's wider than the container itself */
-}
-
-
-.patreon-button {
-    box-shadow: 1px 1px 0px rgba(0, 0, 0, 0.2);
-    line-height: 36px !important;
-    min-width: 190px;
-    max-width: 190px;
-    display: flex !important; /* Changed to flex and kept !important */
-    align-items: center; /* Ensure items are vertically centered */
-    background-color: #6366f1;
-    padding-left: 0.5rem;
-    padding-right: 0.5rem;
-    text-align: center !important;
-    border-radius: 7px;
-    color: #fff;
-    cursor: pointer;
-    overflow-wrap: break-word;
-    vertical-align: middle;
-    border: 0 none #fff !important;
-    font-family:
-        "Quicksand",
-        Helvetica,
-        Century Gothic,
-        sans-serif !important;
-    text-decoration: none;
-    text-shadow: none;
-    font-weight: 700 !important;
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-}
-
-.patreon-button:hover {
-    opacity: 0.85;
-    color: #f5f5f5 !important;
-    text-decoration: none !important;
-}
