--- web/solid/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-        "lib": ["ES2021.String"]
-    },
-    "files": [],
-    "references": [
-        {
-            "path": "./tsconfig.app.json"
-        },
-        {
-            "path": "./tsconfig.node.json"
-        }
-    ]
-}
