--- web/src/app.css
@@ -1,7 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-* {
-	min-width: 0;
-}
