+++ web/tailwind.config.js
@@ -1,8 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-	content: ["./src/**/*.{html,js,svelte,ts}"],
-	theme: {
-		extend: {},
-	},
-	plugins: [],
+    content: ["./src/**/*.{html,js,jsx,ts,tsx}"],
+    theme: {
+        extend: {},
+    },
+    plugins: [],
 };
